+++ lib/node/model/TransitionPath.d.ts
@@ -1,6 +1,8 @@
 import { NamedElement } from './NamedElement';
+import { TransitionKind } from './TransitionKind';
 /** Interface describing elements to leave and enter when traversing the transition; derived from the source and target using the TransitionType strategy. */
 export declare class TransitionPath {
+    readonly kind: TransitionKind;
     readonly leave: NamedElement | undefined;
     readonly enter: Array<NamedElement> | undefined;
     /**
@@ -8,5 +10,5 @@ export declare class TransitionPath {
      * @param leave The optional named element to leave when traversing a transition.
      * @param enter The optional set of elements to enter when traversing a transition.
      */
-    constructor(leave?: NamedElement | undefined, enter?: Array<NamedElement> | undefined);
+    constructor(kind: TransitionKind, leave?: NamedElement | undefined, enter?: Array<NamedElement> | undefined);
 }